requirements.txt
pyserial==3.4
PyQt5==5.11.2
python-gitlab==1.5.1
